Generate model classes for all geometry defined by Bedrock.
One more step towards getting rid of bedrock dependency

+++ src/Tools/ResourceConverterCore/Converter/ResourceConverter.cs
@@ -3,8 +3,9 @@ using System.Collections.Generic;
 using System.IO;
 using System.Text;
 using Alex.ResourcePackLib.Json.Models.Entities;
+using Newtonsoft.Json;
 using NLog;
-using ResourceConverter.Templates;
+using Templates;
 
 namespace ResourceConverter
 {
@@ -39,22 +40,77 @@ namespace ResourceConverter
             int totalCount = loader.EntityModels.Count;
             foreach (var model in loader.EntityModels)
             {
+	            var pct = 100D * ((double)count / (double)totalCount);
+
                 Log.Info($"Starting Template Processing for '{model.Key}'");
 
                 //template.Session["CurrentModelName"] = model.Key;
                 //template.Session["CurrentModel"] = model.Value;
-                ResourceConverterContext.CurrentModelName = model.Key;
+                ResourceConverterContext.CurrentModelName = CodeTypeName(model.Value.Name);
                 ResourceConverterContext.CurrentModel = model.Value;
 
-
                 var output = template.TransformText();
-                var outputPath = Path.Combine(outputDirectoryPath, model.Key + ".cs");
+                var outputPath = Path.Combine(outputDirectoryPath, CodeTypeName(model.Value.Name) + "Model.cs");
+                if (File.Exists(outputPath))
+                {
+					Log.Warn($"Class already exists: {outputPath} ({count}/{totalCount}) - {pct:F1}%");
+                }
+                else
+                {
+	                File.WriteAllText(outputPath, output);
+	                Log.Info($"Successfully Processed Template for entity '{model.Key}' ({count}/{totalCount}) - {pct:F1}%");
+                }
 
-                File.WriteAllText(outputPath, output);
                 count++;
-                var pct = count / (double)totalCount;
-                Log.Info($"Successfully Processed Template for entity '{model.Key}' ({count}/{totalCount}) - {pct:F1}%");
+                // Log.Info($"Successfully Processed Template for entity '{model.Key}' ({count}/{totalCount}) - {pct:F1}%");
             }
+        }
+
+        private static string CodeTypeName(string name)
+        {
+	        if (name.StartsWith("geometry."))
+	        {
+		        name = name.Substring("geometry.".Length);
+
+	        }
+	        else if (name.StartsWith("definition."))
+	        {
+		        name = name.Substring("definition.".Length);
+            }
+
+	        return CodeName(name, true);
+        }
+
+        private static string CodeName(string name, bool firstUpper = false)
+        {
+	        name = name.ToLowerInvariant();
+
+	        string result = name;
+	        bool upperCase = firstUpper;
+
+	        result = string.Empty;
+	        for (int i = 0; i < name.Length; i++)
+	        {
+		        if (name[i] == '.' || name[i] == ' ' || name[i] == '_')
+		        {
+			        upperCase = true;
+		        }
+		        else
+		        {
+			        if ((i == 0 && firstUpper) || upperCase)
+			        {
+				        result += name[i].ToString().ToUpperInvariant();
+				        upperCase = false;
+			        }
+			        else
+			        {
+				        result += name[i];
+			        }
+		        }
+	        }
+
+	        result = result.Replace(@"[]", "s");
+	        return result;
         }
 
         private static string ToPascalCase(string key)
